Move setupWindow away from main since its shared
